+++ src/fio/FioEngine.ts
@@ -1296,6 +1296,7 @@ export class FioEngine extends CurrencyEngine<FioTools, SafeFioWalletInfo> {
 
   // This routine is called once a wallet needs to start querying the network
   async startEngine(): Promise<void> {
+    await this.waitForLoaded()
     this.addToLoop('checkBlockchainInnerLoop', BLOCKCHAIN_POLL_MILLISECONDS)
     this.addToLoop('checkAccountInnerLoop', ADDRESS_POLL_MILLISECONDS)
     this.addToLoop('checkTransactionsInnerLoop', TRANSACTION_POLL_MILLISECONDS)