+++ src/stellar/StellarEngine.ts
@@ -449,8 +449,8 @@ export class StellarEngine extends CurrencyEngine<
   // Public methods
   // ****************************************************************************
 
-  // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-  async startEngine() {
+  async startEngine(): Promise<void> {
+    await this.waitForLoaded()
     this.addToLoop('queryFee', BLOCKCHAIN_POLL_MILLISECONDS)
     this.addToLoop('checkBlockchainInnerLoop', BLOCKCHAIN_POLL_MILLISECONDS)
     this.addToLoop('checkAccountInnerLoop', ADDRESS_POLL_MILLISECONDS)