+++ src/ton/TonEngine.ts
@@ -236,6 +236,7 @@ export class TonEngine extends CurrencyEngine<TonTools, SafeCommonWalletInfo> {
   // // ****************************************************************************
 
   async startEngine(): Promise<void> {
+    await this.waitForLoaded()
     this.addToLoop('queryBalance', ADDRESS_POLL_MILLISECONDS)
     this.addToLoop('queryTransactions', ADDRESS_POLL_MILLISECONDS)
     await super.startEngine()