+++ src/ethereum/info/filecoinFevmInfo.ts
@@ -41,11 +41,11 @@ const networkFees: EthereumFees = {
     },
     gasPrice: {
       lowFee: '1000000001',
-      standardFeeLow: '40000000001',
-      standardFeeHigh: '300000000001',
+      standardFeeLow: '3000000001',
+      standardFeeHigh: '15000000001',
       standardFeeLowAmount: '100000000000000000',
       standardFeeHighAmount: '10000000000000000000',
-      highFee: '40000000001',
+      highFee: '30000000001',
       minGasPrice: '1000000000'
     },
     minPriorityFee: undefined