+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 4.48.2 (2025-06-05)
+
 - changed: (Hedera) Save memo to queried transactions
 - fixed: (Hedera) Fix missing transactions by specifying query limit
 - fixed: Updated FEVM default gas price settings.